Change pre-generated randome sampled dataset
