--- bert_pytorch/build_vocab.py
@@ -1,19 +1,0 @@
-import argparse
-
-from bert_pytorch.dataset import WordVocab
-
-
-def build():
-    parser = argparse.ArgumentParser()
-    parser.add_argument("-c", "--corpus_path", required=True, type=str)
-    parser.add_argument("-o", "--output_path", required=True, type=str)
-    parser.add_argument("-s", "--vocab_size", type=int, default=None)
-    parser.add_argument("-e", "--encoding", type=str, default="utf-8")
-    parser.add_argument("-m", "--min_freq", type=int, default=1)
-    args = parser.parse_args()
-
-    with open(args.corpus_path, "r", encoding=args.encoding) as f:
-        vocab = WordVocab(f, max_size=args.vocab_size, min_freq=args.min_freq)
-
-    print("VOCAB SIZE:", len(vocab))
-    vocab.save_vocab(args.output_path)
